% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triadR.R
\name{triadR}
\alias{triadR}
\title{triadR: An R package for estimating the Triadic Relations Model
from triadic data with generalized round-robin designs}
\description{
Analysize generalized round-robin and
}
\author{
Man-Nok Wong \email{mannok@mnwong.com}

David A. Kenny \email{david.kenny@uconn.edu}
}
