% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trm.R
\name{trm}
\alias{trm}
\title{triadR: Analyzing Generalized Round-Robin Data Using R}
\usage{
trm(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame in long format}
}
\description{
The function trm() performs Tradic Relations Model (TRM) analyses for single or multiple generalized round-robin groups.
}
\examples{
# Load the package
library(triadR)

# Univariate Analysis
data(PM1997)
trm.1 <- trm(var ~ jid*aid*pid, PM1997)
}
